# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mini-EMR API",
    "version": "1.0.0",
    "description": "API Documentation for the Mini-EMR system."
  },
  "paths": {
    "/api/prescriptions": {
      "get": {
        "summary": "Get all prescriptions",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Prescriptions"
        ]
      },
      "post": {
        "summary": "Create a prescription",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "medication": {
                    "type": "string"
                  },
                  "dosage": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "integer"
                  },
                  "refill_on": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "refill_schedule": {
                    "type": "string"
                  },
                  "userId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Prescriptions"
        ]
      }
    },
    "/api/prescriptions/{id}": {
      "get": {
        "summary": "Get prescription details by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "403": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "tags": [
          "Prescriptions"
        ]
      },
      "patch": {
        "summary": "Update a prescription",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "medication": {
                    "type": "string"
                  },
                  "dosage": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "integer"
                  },
                  "refill_on": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "refill_schedule": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Prescriptions"
        ]
      },
      "delete": {
        "summary": "Delete a prescription",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Prescriptions"
        ]
      }
    },
    "/api/appointments": {
      "get": {
        "summary": "Get all appointments",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Appointments"
        ]
      },
      "post": {
        "summary": "Create an appointment",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "provider": {
                    "type": "string"
                  },
                  "datetime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "repeat": {
                    "type": "string"
                  },
                  "userId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Appointments"
        ]
      }
    },
    "/api/appointments/{id}": {
      "get": {
        "summary": "Get appointment details by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "403": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "tags": [
          "Appointments"
        ]
      },
      "patch": {
        "summary": "Update an appointment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "provider": {
                    "type": "string"
                  },
                  "datetime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "repeat": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Appointments"
        ]
      },
      "delete": {
        "summary": "Delete an appointment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Appointments"
        ]
      }
    },
    "/api/patients": {
      "get": {
        "summary": "Get all patients",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Patients"
        ]
      },
      "post": {
        "summary": "Create a patient",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Patients"
        ]
      }
    },
    "/api/patients/{id}": {
      "get": {
        "summary": "Get patient details by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "403": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "tags": [
          "Patients"
        ]
      },
      "patch": {
        "summary": "Update a patient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Patients"
        ]
      },
      "delete": {
        "summary": "Delete a patient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Patients"
        ]
      }
    },
    "/api/auth/signup": {
      "post": {
        "summary": "Patient Signup",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success"
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Patient Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "summary": "Refresh Access Token",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}